structure admin a bit better
