+++ clients/admin/src/main.rs
@@ -1,15 +1,14 @@
+mod account;
+mod disappear;
 mod error;
+mod validate;
 
 use std::env;
 
 use structopt::StructOpt;
 
-use dialoguer::theme::ColorfulTheme;
-use dialoguer::Confirm;
-
 use crate::error::Error;
-use lockbook_core::{base64, Core};
-use lockbook_core::{AccountFilter, AccountIdentifier, Config, PublicKey, Uuid};
+use lockbook_core::{Config, Core, Uuid};
 
 #[derive(Debug, PartialEq, Eq, StructOpt)]
 pub enum Admin {
@@ -27,6 +26,9 @@ pub enum Admin {
 
     /// Validates file trees of all users on the server and prints any failures
     ValidateAccount { username: String },
+
+    /// Performs server-wide integrity checks
+    ValidateServer,
 
     /// List all users
     ListUsers {
@@ -64,110 +66,17 @@ pub fn main() {
     let core = Core::init(&Config { writeable_path, logs: true, colored_logs: true }).unwrap();
 
     let result = match Admin::from_args() {
-        Admin::DisappearAccount { username } => disappear_account(&core, username),
-        Admin::DisappearFile { id } => disappear_file(&core, id),
-        Admin::ValidateAccount { username } => server_validate(&core, username),
+        Admin::DisappearAccount { username } => disappear::account(&core, username),
         Admin::ListUsers { premium, google_play_premium, stripe_premium } => {
-            list_users(&core, premium, google_play_premium, stripe_premium)
+            account::list(&core, premium, google_play_premium, stripe_premium)
         }
-        Admin::AccountInfo { username, public_key } => account_info(&core, username, public_key),
+        Admin::AccountInfo { username, public_key } => account::info(&core, username, public_key),
+        Admin::DisappearFile { id } => disappear::file(&core, id),
+        Admin::ValidateAccount { username } => validate::account(&core, username),
+        Admin::ValidateServer => todo!(),
     };
 
     if result.is_err() {
         panic!("unsuccessful completion")
     }
-}
-
-fn disappear_file(core: &Core, id: Uuid) -> Res<()> {
-    let maybe_confirm = Confirm::with_theme(&ColorfulTheme::default())
-        .with_prompt(format!("Are you sure you want to delete '{}'?", id))
-        .interact_opt()?;
-
-    if maybe_confirm.unwrap_or(false) {
-        core.admin_disappear_file(id)?;
-    }
-    Ok(())
-}
-
-fn disappear_account(core: &Core, username: String) -> Res<()> {
-    let maybe_confirm = Confirm::with_theme(&ColorfulTheme::default())
-        .with_prompt(format!("Are you sure you want to delete '{}'?", username))
-        .interact_opt()?;
-
-    if maybe_confirm.unwrap_or(false) {
-        core.admin_disappear_account(&username)?;
-
-        println!("Account deleted!");
-    }
-
-    Ok(())
-}
-
-fn list_users(
-    core: &Core, premium: bool, google_play_premium: bool, stripe_premium: bool,
-) -> Res<()> {
-    let filter = if premium {
-        Some(AccountFilter::Premium)
-    } else if google_play_premium {
-        Some(AccountFilter::GooglePlayPremium)
-    } else if stripe_premium {
-        Some(AccountFilter::StripePremium)
-    } else {
-        None
-    };
-
-    let users = core.admin_list_users(filter.clone())?;
-
-    if users.is_empty() {
-        let msg = match filter {
-            None => "There are no users.",
-            Some(AccountFilter::Premium) => "There are no premium users.",
-            Some(AccountFilter::GooglePlayPremium) => "There are no premium google play users.",
-            Some(AccountFilter::StripePremium) => "There are no premium stripe users.",
-        };
-
-        println!("{}", msg);
-    } else {
-        for user in users {
-            println!("{}", user);
-        }
-    }
-
-    Ok(())
-}
-
-fn account_info(core: &Core, username: Option<String>, public_key: Option<String>) -> Res<()> {
-    let identifier = if let Some(username) = username {
-        AccountIdentifier::Username(username)
-    } else if let Some(public_key) = public_key {
-        AccountIdentifier::PublicKey(PublicKey::parse_compressed(<&[u8; 33]>::try_from(
-            base64::decode(public_key)?.as_slice(),
-        )?)?)
-    } else {
-        println!("Please specify a username or public key.");
-        return Err(Error);
-    };
-
-    let account_info = core.admin_get_account_info(identifier)?;
-
-    println!("{:#?}", account_info);
-
-    Ok(())
-}
-
-fn server_validate(core: &Core, username: String) -> Res<()> {
-    println!("Validating server...");
-
-    let validation_failures = core.admin_server_validate(&username)?;
-    for failure in validation_failures.tree_validation_failures {
-        println!("tree validation failure: {:?}", failure);
-    }
-    for failure in validation_failures.documents_missing_content {
-        println!("document missing content: {:?}", failure);
-    }
-    for failure in validation_failures.documents_missing_size {
-        println!("document missing size: {:?}", failure);
-    }
-
-    Ok(())
 }
